backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==1.4.0
psycopg2-binary==2.9.5
requests==2.28.1
python-dotenv
passlib[bcrypt]
pyjwt
requests
email-validator
html_table_parser
selenium

